refactor tutorial
