+++ just_jaguars/main.py
@@ -6,6 +6,10 @@ from assets import (
     geothermal, biomass, nuclear,
     hydrogen, ghg_capture_tech)  # Importing the images for the icons
 
+# tutorial:
+import tutorial
+tutorial.__dict__  # because flake8 ;-;
+
 # Details about the window:
 window_width = 1152
 aspect_ratio = 9 / 16
@@ -364,41 +368,6 @@ def fps_counter():
         fps_displayed = str(int(clock.get_fps()))
         fps_text = arial.render(f'FPS: {fps_displayed}/60', False, (0, 0, 0))
     window.blit(fps_text, (0, window_height - 20))
-
-
-def tutorial():
-    """Credit goes to https://docs.opencv.org/master/dd/d43/tutorial_py_video_display.html"""
-    import cv2
-
-    cap = cv2.VideoCapture('assets/tutorial.mov')
-
-    while(cap.isOpened()):
-        ret, frame = cap.read()
-
-        if not ret or cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-
-        cv2.imshow('frame', frame)
-
-    cap.release()
-    cv2.destroyAllWindows()
-
-# test if the tutorial needs to be played
-
-
-try:
-    tutorial_file = open('saves/tutorial.txt')  # error causing function 1
-
-    tutorial_content = tutorial_file.read()
-    tutorial_file.close()
-    del tutorial_file
-
-    assert tutorial_content == 'done'  # error causing function 2
-
-except Exception:
-    tutorial()
-    with open('saves/tutorial.txt', 'w') as f:
-        f.write('done')
 
 
 pause = False
